refactor(apps/web): remove legacy playback sync files and introduce audio metadata utilities
- Deleted outdated `playback-sync.ts` and `playback-sync.test.ts` files to streamline the codebase.
- Added new utilities for extracting audio metadata and cover art from audio files, enhancing functionality for audio processing.
- Introduced tests for audio metadata extraction and filename cleaning, ensuring robust handling of audio file metadata.
- Implemented playback device management utilities to improve device identification and management within the playback context.
