+++ apps/web/src/lib/playback/sync/playback-sync.commands.test.ts
@@ -1,0 +1,364 @@
+import { createPlayerStateMock } from '@/components/app/test-utils/player-test-utils';
+import { usePlayerStore } from '@/stores/player-store/player.store';
+import type { ListPlaybackDeviceEntry } from '@repo/contracts';
+import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
+import { createPlaybackSocket } from '../playback-socket';
+import {
+  connectPlaybackSync,
+  disconnectPlaybackSync,
+  emitCurrentTimeSync,
+  listPlaybackDevices,
+  setActivePlaybackDevice,
+  syncPlayingStateToServer,
+} from './playback-sync';
+import {
+  asSocketMock,
+  basePlaybackSyncTestState,
+  createPlaybackSocketMock,
+  flushMicrotasks,
+  playbackStateFixture,
+  type EmitCallbackPayload,
+  type PlaybackSocketMock,
+} from './playback-sync.test-helpers';
+
+vi.mock('../playback-socket', () => ({
+  createPlaybackSocket: vi.fn(() => ({
+    on: vi.fn(),
+    emit: vi.fn(),
+    disconnect: vi.fn(),
+    removeAllListeners: vi.fn(),
+    connected: true,
+  })),
+}));
+
+vi.mock('../playback-device', () => ({
+  getLocalPlaybackDeviceMetadata: vi.fn(() => ({
+    playbackDeviceId: 'device-1',
+    deviceName: 'Web',
+    deviceIcon: 'desktop',
+  })),
+}));
+
+vi.mock('@/stores/player-store/player.store', () => ({
+  usePlayerStore: {
+    getState: vi.fn(),
+    setState: vi.fn(),
+  },
+}));
+
+describe('playback-sync commands', () => {
+  let mockSocket: PlaybackSocketMock;
+
+  beforeEach(() => {
+    vi.clearAllMocks();
+    mockSocket = createPlaybackSocketMock();
+    vi.mocked(createPlaybackSocket).mockReturnValue(asSocketMock(mockSocket));
+    vi.mocked(usePlayerStore.getState).mockReturnValue({ ...basePlaybackSyncTestState });
+  });
+
+  afterEach(() => {
+    disconnectPlaybackSync();
+  });
+
+  describe('syncPlayingStateToServer', () => {
+    it('returns early when connected but there is no current track', async () => {
+      vi.mocked(usePlayerStore.getState).mockReturnValue(
+        createPlayerStateMock({ ...basePlaybackSyncTestState, currentTrack: null }),
+      );
+      connectPlaybackSync('token');
+      syncPlayingStateToServer(false);
+      await flushMicrotasks();
+
+      expect(mockSocket.emit).not.toHaveBeenCalledWith(
+        'command:set-playing-state',
+        expect.any(Object),
+        expect.any(Function),
+      );
+    });
+  });
+
+  describe('listPlaybackDevices', () => {
+    it('emits list-devices and updates store', async () => {
+      connectPlaybackSync('token');
+      const listDevice: ListPlaybackDeviceEntry = {
+        id: 'd1',
+        name: 'D1',
+        icon: 'desktop',
+        isActive: true,
+        isCurrentDevice: false,
+      };
+      mockSocket.emit.mockImplementation(
+        (
+          event: string,
+          _data: unknown,
+          callback: (r: { devices: ListPlaybackDeviceEntry[] }) => void,
+        ) => {
+          if (event === 'query:list-devices') {
+            callback({ devices: [listDevice] });
+          }
+        },
+      );
+
+      await listPlaybackDevices();
+      expect(usePlayerStore.getState().setPlaybackDevices).toHaveBeenCalledWith([listDevice]);
+    });
+
+    it('returns early if disconnected', async () => {
+      mockSocket.connected = false;
+      await listPlaybackDevices();
+      expect(mockSocket.emit).not.toHaveBeenCalledWith(
+        'query:list-devices',
+        expect.any(Object),
+        expect.any(Function),
+      );
+    });
+  });
+
+  describe('emitCurrentTimeSync', () => {
+    it('emits current time and handles response', async () => {
+      connectPlaybackSync('token');
+      mockSocket.emit.mockImplementation(
+        (event: string, _data: unknown, callback: (r: EmitCallbackPayload) => void) => {
+          if (event === 'command:set-current-time-state') {
+            callback({ currentTime: 50, version: 10 });
+          }
+        },
+      );
+
+      await emitCurrentTimeSync(50);
+      expect(mockSocket.emit).toHaveBeenCalledWith(
+        'command:set-current-time-state',
+        expect.objectContaining({ currentTime: 50 }),
+        expect.any(Function),
+      );
+      expect(usePlayerStore.setState).toHaveBeenCalledWith(
+        expect.objectContaining({ currentTime: 50, playbackVersion: 10 }),
+      );
+    });
+
+    it('handles version conflict by refetching state', async () => {
+      connectPlaybackSync('token');
+      mockSocket.emit.mockImplementation(
+        (event: string, _data: unknown, callback: (r: EmitCallbackPayload) => void) => {
+          if (event === 'command:set-current-time-state') {
+            callback({ error: 'version mismatch', code: 'CONFLICT' });
+          } else if (event === 'query:get-state') {
+            callback(playbackStateFixture({ version: 11 }));
+          }
+        },
+      );
+
+      await emitCurrentTimeSync(50);
+      expect(mockSocket.emit).toHaveBeenCalledWith('query:get-state', {}, expect.any(Function));
+      expect(usePlayerStore.getState().applyPlaybackStateFromServer).toHaveBeenCalled();
+    });
+
+    it('isVersionConflict handles various message formats', async () => {
+      connectPlaybackSync('token');
+      mockSocket.emit.mockImplementation(
+        (event: string, _data: unknown, callback: (r: EmitCallbackPayload) => void) => {
+          if (event === 'command:set-current-time-state') {
+            callback({ error: 'version mismatch', code: 'OTHER' });
+          } else if (event === 'query:get-state') {
+            callback(playbackStateFixture({ version: 100 }));
+          }
+        },
+      );
+      await emitCurrentTimeSync(10);
+      expect(mockSocket.emit).toHaveBeenCalledWith('query:get-state', {}, expect.any(Function));
+    });
+
+    it('isVersionConflict returns false when not a conflict', async () => {
+      connectPlaybackSync('token');
+      mockSocket.emit.mockImplementation(
+        (_e: string, _d: unknown, callback: (r: EmitCallbackPayload) => void) => {
+          callback({ error: 'general error', code: 'ERROR' });
+        },
+      );
+      await emitCurrentTimeSync(10);
+      expect(mockSocket.emit).not.toHaveBeenCalledWith(
+        'query:get-state',
+        expect.any(Object),
+        expect.any(Function),
+      );
+    });
+
+    it('returns early if disconnected or invalid state', async () => {
+      await emitCurrentTimeSync(10);
+      expect(mockSocket.emit).not.toHaveBeenCalledWith(
+        'command:set-current-time-state',
+        expect.any(Object),
+        expect.any(Function),
+      );
+
+      connectPlaybackSync('token');
+
+      vi.mocked(usePlayerStore.getState).mockReturnValue(
+        createPlayerStateMock({ ...basePlaybackSyncTestState, currentTrack: null }),
+      );
+      await emitCurrentTimeSync(10);
+      expect(mockSocket.emit).not.toHaveBeenCalledWith(
+        'command:set-current-time-state',
+        expect.any(Object),
+        expect.any(Function),
+      );
+
+      vi.mocked(usePlayerStore.getState).mockReturnValue(
+        createPlayerStateMock({ ...basePlaybackSyncTestState, playbackVersion: 0 }),
+      );
+      await emitCurrentTimeSync(10);
+      expect(mockSocket.emit).not.toHaveBeenCalledWith(
+        'command:set-current-time-state',
+        expect.any(Object),
+        expect.any(Function),
+      );
+    });
+
+    it('handles null resync payload and non-error falsy result', async () => {
+      connectPlaybackSync('token');
+      mockSocket.emit
+        .mockImplementationOnce(
+          (_e: string, _d: unknown, callback: (r: EmitCallbackPayload) => void) => {
+            callback({ error: 'conflict', code: 'CONFLICT' });
+          },
+        )
+        .mockImplementationOnce(
+          (_e: string, _d: unknown, callback: (r: EmitCallbackPayload) => void) => {
+            callback(null);
+          },
+        );
+      await emitCurrentTimeSync(10);
+      expect(usePlayerStore.getState().applyPlaybackStateFromServer).not.toHaveBeenCalled();
+
+      mockSocket.emit.mockImplementationOnce(
+        (_e: string, _d: unknown, callback: (r: EmitCallbackPayload) => void) => {
+          callback(null);
+        },
+      );
+      await emitCurrentTimeSync(10);
+      expect(usePlayerStore.setState).not.toHaveBeenCalled();
+    });
+  });
+
+  describe('setActivePlaybackDevice', () => {
+    it('emits set-active-device', async () => {
+      connectPlaybackSync('token');
+      mockSocket.emit.mockImplementation(
+        (event: string, _data: unknown, callback: (r: EmitCallbackPayload) => void) => {
+          if (event === 'command:set-active-device') {
+            callback(playbackStateFixture({ version: 20 }));
+          } else if (event === 'query:list-devices') {
+            callback({ devices: [] });
+          }
+        },
+      );
+
+      await setActivePlaybackDevice('new-device');
+      expect(mockSocket.emit).toHaveBeenCalledWith(
+        'command:set-active-device',
+        expect.objectContaining({ deviceId: 'new-device' }),
+        expect.any(Function),
+      );
+      expect(mockSocket.emit).toHaveBeenCalledWith('query:list-devices', {}, expect.any(Function));
+    });
+
+    it('handles no playbackVersion', async () => {
+      vi.mocked(usePlayerStore.getState).mockReturnValue(
+        createPlayerStateMock({ ...basePlaybackSyncTestState, playbackVersion: 0 }),
+      );
+      connectPlaybackSync('token');
+
+      await setActivePlaybackDevice('new-device');
+      expect(mockSocket.emit).not.toHaveBeenCalledWith(
+        'command:set-active-device',
+        expect.any(Object),
+        expect.any(Function),
+      );
+    });
+
+    it('returns early if disconnected', async () => {
+      mockSocket.connected = false;
+      await setActivePlaybackDevice('d1');
+      expect(mockSocket.emit).not.toHaveBeenCalledWith(
+        'command:set-active-device',
+        expect.any(Object),
+        expect.any(Function),
+      );
+    });
+
+    it('on version conflict refetches state via get-state', async () => {
+      connectPlaybackSync('token');
+      mockSocket.emit.mockImplementation(
+        (event: string, _data: unknown, callback: (r: EmitCallbackPayload) => void) => {
+          if (event === 'command:set-active-device') {
+            callback({ error: 'version mismatch', code: 'CONFLICT' });
+          } else if (event === 'query:get-state') {
+            callback(playbackStateFixture({ version: 30 }));
+          }
+        },
+      );
+
+      await setActivePlaybackDevice('new-device');
+
+      expect(mockSocket.emit).toHaveBeenCalledWith('query:get-state', {}, expect.any(Function));
+      expect(usePlayerStore.getState().applyPlaybackStateFromServer).toHaveBeenCalledWith(
+        expect.objectContaining({ version: 30 }),
+      );
+    });
+
+    it('on non-conflict error resolves without get-state', async () => {
+      connectPlaybackSync('token');
+      mockSocket.emit.mockImplementation(
+        (_event: string, _data: unknown, callback: (r: EmitCallbackPayload) => void) => {
+          callback({ error: 'other', code: 'ERROR' });
+        },
+      );
+
+      await setActivePlaybackDevice('new-device');
+
+      expect(mockSocket.emit).not.toHaveBeenCalledWith(
+        'query:get-state',
+        expect.any(Object),
+        expect.any(Function),
+      );
+    });
+
+    it('on conflict when get-state returns null, does not apply server state', async () => {
+      const apply = vi.mocked(usePlayerStore.getState().applyPlaybackStateFromServer);
+      const callsBefore = apply.mock.calls.length;
+
+      connectPlaybackSync('token');
+      mockSocket.emit.mockImplementation(
+        (event: string, _data: unknown, callback: (r: EmitCallbackPayload) => void) => {
+          if (event === 'command:set-active-device') {
+            callback({ error: 'version mismatch', code: 'CONFLICT' });
+          } else if (event === 'query:get-state') {
+            callback(null);
+          }
+        },
+      );
+
+      await setActivePlaybackDevice('new-device');
+
+      expect(mockSocket.emit).toHaveBeenCalledWith('query:get-state', {}, expect.any(Function));
+      expect(apply.mock.calls.length).toBe(callsBefore);
+    });
+
+    it('when ack is not an error but not a playback state, skips apply and listPlaybackDevices', async () => {
+      const apply = vi.mocked(usePlayerStore.getState().applyPlaybackStateFromServer);
+
+      connectPlaybackSync('token');
+      const callsBefore = apply.mock.calls.length;
+
+      mockSocket.emit.mockImplementation(
+        (_event: string, _data: unknown, callback: (r: EmitCallbackPayload) => void) => {
+          callback({});
+        },
+      );
+
+      await setActivePlaybackDevice('new-device');
+
+      expect(apply.mock.calls.length).toBe(callsBefore);
+    });
+  });
+});
