+++ apps/web/src/lib/playback/sync/playback-sync.ts
@@ -1,0 +1,16 @@
+export {
+  afterLocalPlaybackMutation,
+  afterLocalPlaybackMutationWithClaim,
+  emitCurrentTimeSync,
+  listPlaybackDevices,
+  setActivePlaybackDevice,
+  syncPlayingStateToServer
+} from './playback-sync.commands';
+export {
+  connectPlaybackSync,
+  disconnectPlaybackSync,
+  getPlaybackSocket,
+  isPlaybackSyncConnected
+} from './playback-sync.connection';
+export { applyStateFromServer } from './playback-sync.store-bridge';
+
